+++ changelog.md
@@ -18,6 +18,9 @@ You can follow the project also consulting the [commit history](https://github.c
 ## Changelog
 Versions are listed on reverse order, the first is the last one.
 
+### eggs-8.0.15
+* krill finally support changing keyboard
+
 ### eggs-8.0.14
 * re added rsync, after unsquashfs, to let modifications in live to be reflected in the installed system